Update app dependencies in metadata file
Recently the dependency section from the apps metadata was changed
to allow mutually exclusive dependencies [1], in this case either
depend on the rook-ceph or integ-apps application.

This review changes the dependency section in the STX-O
metadata, removing the deprecated "after" key and adding the
new "dependent_apps" section, to allow the dependency to either
integ-apps or rook-ceph applications.

Test Plan:
PASS - Build stx-openstack-helm-fluxcd package and STX-O tarball
PASS - Upload and apply STX-Openstack (with temporary workaround [2])
PASS - Host lock/unlock

[1] - https://review.opendev.org/c/starlingx/config/+/953638
[2] - Currently there is a bug (#2120478) that is being worked on, where
host-lock is impossible after applying STX-O, a workaround for that
is to delete the app from the dependencies that is in an uploaded
state.

Story: 2011351
Task: 52715

Partial-Bug: #2116263
Partial-Bug: #2120478

Change-Id: Iaf94b827a8f50f349fa01d439449f2bd513e9d60
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>