Remove semantic check to allow k8s upgrade
This semantic check is preventing the starlingx to upgrade
kubernetes version with stx-openstack applied following the
instructions on [1].

This check was created in [2] to prevent the app auto-update,
however,this is no longer necessary, as the app will not be
updated without the bundle [3].

By default, stx-openstack should  not be auto-updated on k8s
upgrade. Current logic already ensures that. Sysinv uses
get_app_bundle_for_update"[4] and simply skips the update if
no bundle is found. For stx-openstack upgrade, please follow [5].



Test plan:

PASS: Apply custom stx-openstack to AIO-SX
PASS: Upgrade kubernetes version on platform
PASS: Check that the app wont auto-update

Closes-Bug: #2119329
[1] https://docs.starlingx.io/updates/kubernetes/manual-kubernetes-components-upgrade.html

[2] https://review.opendev.org/c/starlingx/openstack-armada-app/+/800824

[3]https://opendev.org/starlingx/config/src/branch/master/sysinv/sysinv/sysinv/sysinv/conductor/manager.py#L7822

[4]https://opendev.org/starlingx/config/src/branch/master/sysinv/sysinv/sysinv/sysinv/conductor/manager.py#L7822

[5] https://docs.starlingx.io/updates/openstack/index-updates-os-5e0354d270cb.html#software-update


Change-Id: Ied4a80dfb79d667c1a5e40fc994710ca4d31498e
Signed-off-by: Romulo Leite <romulo.leite@windriver.com>