Add Unit Test for Openstack Clients Helm Chart
Since the Clients Helm Chart was created, no related unit tests were
implemented. Because of that, the coverage percentage of ClientsHelm
was below 50%, providing no reliability.

The main method of ClientsHelm is get_overrides, which has some
conditions and depends on other private methods in the class. For this
reason, all conditions for each method were tested.

Bellow are the coverage values before and after the unit tests created:

File				  Prev(%) New(%)
k8sapp_openstack/helm/clients.py  47%	  100%

Another important point about this test is that mocks are not created
for internal methods, but only for instantiated functions. From this
perspective, a strong scenario of atomic tests is established.

TEST PLAN:
- PASS: Unit tests are running and passing
- PASS: Code coverage is at 100%

Story: 2011529
Task: 52816

Change-Id: I2535fe293e580cfca1b333836114036fb179ab40
Signed-off-by: Andre de Held Badur <Andre.deHeldBadur@windriver.com>