Add retry to neutron agents initialization
On Standard deployments with multiple nodes, it was
observed that some neutron agents get into a failure
state with crashLoopBackoff on the stx-openstack install.
This is a very intermittent issue. On a 100 nodes install
there were only  3 compute nodes affected.
The error was happening because the init scrpit of the
agents were not able to get the host name due to a race
condition.

This change adds a retry on getting the hostname of the
running node to avoid getting into the fail state on the
neutron agent pod initialization

Test plan:

- On build system:

PASS: build-pkgs -c -p openstack-helm
PASS: build stx-openstack

- On AIO-SX and Standard installs:

PASS: Apply custom stx-openstack with the change multiple times:
PASS: Check pods health
PASS: Check neutron agents health

Closes-Bug: 2123849

Change-Id: I641fa8f15a487c7edee6a79841f359264919c59b
Signed-off-by: Romulo Leite <romulo.leite@windriver.com>