Fix  neutron openvswitch hostname initialization
It was noticed a openvswitch agent error due to a syntax
error introduced on [1]. The error is caused by a unmatched
parenthesis. This was only on openvswitch part of the patch.
This causes a CrashLoopBackOff error on neutron openvswitch
pod.

Closes-Bug: 2125672

Test plan:

- Build env

PASS: build-pkgs -c -p
PASS: build stx-openstack

- Standard lab

PASS: upload and apply stx-openstack
PASS: check system health
PASS: re-apply the app 3 times

[1] https://review.opendev.org/c/starlingx/openstack-armada-app/+/961372

Change-Id: I928cffc6c2ba1d7c4f467edf9c535a9f80861142
Co-Authored-By: Joao Pereira <joaohenrique.pereira@windriver.com>
Signed-off-by: Romulo Leite <romulo.leite@windriver.com>