Fix fail message when applying ceph/rook configmap
During the pre-apply cycle, the lifecycle plugin will attempt to
detect the available storage backend, falling back to the kube-system
provided configmap.

If Rook Ceph was available but not configured the plugin reported an
unclear error message "Missing ceph-etc config map". Now, the error
message will include additional context specifying that Rook Ceph is
available but not configured including namespace and configmap names
as well an exception.

When trying to configure the fallback kube-system provided storage
backend the plugin could fail with the same message which was
misleading. The correct namespace and configmap resources should now
be reported.

It will, also, show a better message for when the database api is
not available and will throw an exception as well.

Multiple plugins needed additional change to be adapted to the
modify function inside the utils file.

Unit tests were created fo the new and the modified functions.

Test Plan:
PASS - rook-ceph uploaded but not configured, STX-O fails with:
       "Requested backend ceph-rook is not configured"
PASS - rook-ceph applied and ready, STX-O successfuly applies
PASS - rook-ceph is applied but configmap is missing, STX-O fails with:
       "Missing storage backend config map: rook-ceph/ceph-etc"
PASS - platform-integ-apps applied: STX-O successfuly applies
PASS - platform-integ-apps applied but configmap is missing, STX-O fails
       with: "Missing storage backend config map:
              kube-system/rbd-storage-init"

Closes-Bug: #2121944

Change-Id: Ia44806298ba53e339a5555346c3fca84d1c5733a
Signed-off-by: Carlos Henrique <carlos.figueiredofelipe@windriver.com>
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>