Fix mariadb-controller number of replicas
The change [1] alters the way mariadb is deployed. Instead of using
an ingress, now it uses the 'mariadb-controller' pod with 'mariadb'
service. This controller periodically checks if the pods
'mariadb-server' are running and ties the mentioned service to the
first running pod it finds.
However, in a scenario with 2 controllers, if the node hosting
mariadb-controller becomes unreachable, the pod is not deployed
in the other node. Hence, the main service isn't updated and can
be pointing to a pod that is unreachable. This causes all mariadb
queries to fail.

In this change, we set the number of mariadb-controller replicas to be
same number of mariadb-server replicas. By doing this, we make sure that
there will be always a mariadb-controller running, and therefore the
service can be updated even when one of the nodes fails.

References
[1] https://opendev.org/starlingx/openstack-armada-app/commit/b990ebe3b0a86590bb6000e31a4ac217bab7ff50

Test plan
PASS - build python3-k8sapp-openstack package
PASS - generate stx-openstack tarball
PASS - apply stx-openstack in an AIO-DX environment
PASS - verify that there are 2 mariadb-controller pods
PASS - verify that 'mariadb' service is selecting mariadb-server-0 pod
PASS - shutdown the standby node and verify that now the service is
selecting mariadb-server-1 pod
PASS - verify if 'mariadb' service is reachable using wget on the
endpoint

Partial-bug: #2126578

Change-Id: I943a470bca9f12b5a3d87f21ab08d2a8c47ff988
Signed-off-by: Joao Fracarolli <joao.vicentinifracarolli@windriver.com>