Fix ingress-nginx-openstack chart anti-affinities
In an AIO-DX environment, we deploy one ingress for each controller.
However, when one of the controllers is shut down, it causes the
deployment of an additional ingress in the other node, resulting
in 2 ingresses at the same controller. This happens because the
podAntiAffinity keys in ingress's static-overrides don't match any
of the pods' labels.
In this change, we set the correct rules in the podAntiAffinity
field in ingress-nginx-openstack static-overrides.

Test plan:
PASS - Build stx-openstack-helm-fluxcd package
PASS - Buils stx-openstack tarball
PASS - Apply stx-openstack in an AIO-DX environment
PASS - Verify we have one ingress on each controller
PASS - Shutdown one of the controllers and verify we have only
one ingress-nginx-openstack pod running on the other controller

Partial-bug: #2126578

Change-Id: If250e897076c455247d0dedc83a557228f34ed4b
Signed-off-by: Joao Fracarolli <joao.vicentinifracarolli@windriver.com>