Block STX-O remove with deployed servers
Currently there is a known issue that, if STX-O is removed
before deleting all servers in the system, the subsequent
apply will fail to deploy the nova-compute pod. There is a
planned work on to fix this, but in the meantime this review
blocks the app removal if there is any server in the openstack
database.

This review creates a function in utils that returns the output
of the openstack command to list the servers. Furthermore it deals
with the case of trying to remove the application that failed before
the clients pod is deployed. In this case it give a empty string
if the output is None.

It also creates a new lifecycle function to be called in the
semantics check for pre remove os the app. This function is the one
that will effectively block the app removal. The function analyzes
the length of the output to determine if there is or not a VM in the system.

Since the return from the Openstack server list command when no VM
is in the system, is not an empty string, but rather a "\n", it
was necessary to use the strip() function in the return of the
output.

This review also adds the unit tests for the new functions in the
lifecycle and utils file.

Test-Plan:
PASS - Build all packages and STX-O tarball
PASS - Upload and apply STX-Openstack
PASS - App remove is blocked after launching a VM
PASS - App remove after deleting the VM
PASS - App remove if app is in apply-failed state

Story: 2011544
Task: 52871

Change-Id: Idde6ec9a9997dfb917c2b2402a4dff4f909c965c
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>