Add mariadb as a dependency to garbd
On standard deploys, it was observed that fresh stx-openstack
installs were getting a fail state of osh-openstack-garbd pods.
This is happening because the garbd pod needs mariadb running
to be operational.

A solution for that is simply adding the maridb helm chart as a
dependency to the garbd.

Test plan:

PASS: build-pkgs -c -p python-openstackclient
PASS: build stx-openstack
PASS: Apply the app on std system and check pods health
PASS: Reapply stx-openstack
PASS: Remove/Delete the app

Closes-Bug: 2127483

Change-Id: Ifa3a3d7a7b4ba04f8c425d113c16be789e67c829
Signed-off-by: Romulo Leite <romulo.leite@windriver.com>