Add new DEX parameters to keystone values
As part of the initial work on implementing DEX auth
in STX-O, this review adds the new parameters to the
keystone values file.

This new keys and values will be used in future changes
to activate the keystone federation during keystone
deployment, it will also add the possibility to create
minimal user and projects to enable the use of DEX with
STX-O.

For now this change only adds the keys and values, not
changing anything in the deployment itself, thus with this
review DEX can not be activate yet.

With this values the user have 3 options:

 - For the first case (conf.federation.dex_idp.enabled: true),
   enables the option for the wsgi_keystone configuration to
   be modified to include the OIDC settings, and the script
   defined in conf.federation.dex_idp.bootstrap.script will
   be appended to Keystone’s bootstrap script. If desired,
   the user can modify the DEX script commands as they wish.
 - For the second case (conf.federation.dex_idp.bootstrap.enabled: false),
   enables the option for the wsgi_keystone configuration update, but
   nothing will be added to the bootstrap — in this situation,
   the user may choose to manually create groups, projects, roles, etc.
   after applying the chart.
 - Finally, the last case is the default scenario, where nothing
   changes and Keystone is deployed and started as it currently is.

Test Plan:
PASS - Build OSH package and STX-O tarball
PASS - Keystone values contains the new parameters
PASS - Application upload/apply/remove/delete
PASS - Nothing change during app apply (for now)

Story: 2011517
Task: 52957

Change-Id: Icd8480f39ec4dcfb326d937e3b189f2fac9cda7b
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>