Update NetApp configuration for Cinder
This change updates and improves some functions that are used
to configure NetApp storage backend.
First, we modified the check_netapp_backend function to be able to
detect Fiber Channel Protocol (FCP) in addition to NFS and iSCSI.
To do this, a new check was added to get the nasType or sanType fields
from the tridentbackend objects.
We're also using K8s' storageclasses to check if we have trident
classes available.

New fields were added to Cinder's values.yaml and static overrides
to allow for manual selection of the storage backend to be used.
We added support for selecting between the following backends:

- ceph
- netapp-nfs
- netapp-iscsi
- netapp-fc

Also, an array defining the priority of each backend was introduced
so we can select which will be the default backend when we have multiple
backends enabled by the user. Now, if NetApp is the only backend
enabled, the plugin will auto select one of NetApp's backends as
the default one.

Note:
FC will be fully supported on a future change, so in this one we
just added the base code to enable the plugin to detect FC as a
valid option.

TEST PLAN:
All tests were run on the following environment:
Virtual Environment with AIO-SX + NetApp simulator
PASS - Unit tests passed
PASS - build-pkgs -a -l openstack
PASS - build stx-openstack tarball
PASS - application upload, apply and remove
PASS - storage backends tests as described below

Storage backends tests:
The tests were run for each of the following scenarios

+--------------+------------+--------------+----------------+-------+
| Scenario     | NetApp NFS | NetApp iSCSI | Discovery Mode | Notes |
+--------------+------------+--------------+----------------+-------+
| Ceph only    | Disabled   | Disabled     | —              |       |
+--------------+------------+--------------+----------------+-------+
| Ceph + NFS   | Enabled    | Disabled     | Auto           |       |
+--------------+------------+--------------+----------------+-------+
| Ceph + iSCSI | Disabled   | Enabled      | Auto           | 1     |
+--------------+------------+--------------+----------------+-------+
| Ceph + NFS   | Enabled    | Disabled     | Overrides      | 2,3   |
+--------------+------------+--------------+----------------+-------+
| Ceph + iSCSI | Disabled   | Enabled      | Overrides      | 2     |
+--------------+------------+--------------+----------------+-------+

Notes:
(1) For iSCSI driver to load correctly, we need to set the
readonlyRootFilesystem option in security_context to False. Also, to
create a volume, it was necessary to increase the capacity of the NetApp
Flexvol to 2 GB.
(2) To create volumes correctly, we need to have the NetApp configs (IP,
port, credentials) available at /etc/cinder/conf/backends.conf. It's
necessary to implement inside the autodiscovery some code to gather this
info from the K8s resources (tridentbackendsconfig, tridentbackends, secrets)
and override them using the plugin.
(3) For NFS, we need to specify the nfs_shares config. It's still necessary
to study how to gather the volume name using kubectl/tridentctl

For each scenario we validate the following:
PASS - check cinder-volume pod logs for errors
PASS - in cinder-volume pod, check cinder.conf for
'enabled_backends' and 'default_volume_type' configurations.
The values will depend on the enabled backends and on the
priority list:
  PASS - Test with Ceph as the higher priority
  PASS - Test with NetApp as the highest priority
PASS - create and delete volumes with all the backends
under test (see notes 2 and 3)

Story: 2011281
Task: 52739

Change-Id: I450faefe96d00f40f8f989cb635c11a9c2000651
Signed-off-by: Joao Fracarolli <joao.vicentinifracarolli@windriver.com>