Add DEX integration to keystone bootstrap
In a previous review [1], new parameters were added
to the keystone values file, related to the DEX integration
with STX Openstack. That review was the first on a series
of changes to allow STX Openstack users to run the application
with DEX.

On this review, the new patch is incremeted adding the bootstrap
script to the keystone bootstrap job. This script can be customized
as per the user needs. The commands being executed by default are
minimal commands to create user, project, etc. to a simple dex usage
with STX-O.

 - If the Values.conf.federation.dex_idp.enabled = true: then, the
   dex_mapping.json will be created.
 - If Values.conf.federation.dex_idp.enabled = true and
   Values.conf.federation.dex_idp.bootstrap.enabled = true: then,
   the dex-bootstrap.sh will be created and executed after the
   keystone bootstrap commands.
 - If Values.conf.federation.dex_idp.enabled = false: then,
   keystone is deployed as it is currently.

[1] - https://review.opendev.org/c/starlingx/openstack-armada-app/+/964701

Test Plan:
PASS - Build OSH package and STX-O tarball
PASS - Upload and apply with default keystone config
PASS - Only dex_idp enabled: dex_mappings.json is created
       inside keystone configmap, but no command is executed
PASS - dex_idp.bootstrap also enabled: dex-bootstrap.sh is also
       created in the configmap and the commands from the dex
       bootstrap script are executed in the keystone bootstrap job
PASS - Any change in default dex config is reflected in the dex
       mappings and/or bootstrap commands.

Story: 2011517
Task: 52958

Change-Id: I3c24314133613877ec30837ba8b18ff1420c1021
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>