Add dex URL override to Keystone plugin
As part of the work to integrate STX OpenStack and DEX,
the issuer_url parameter configured in the system needs
to be acquired by Keystone for internal configuration.

This change adds a logic to retrieve the issuer_url from
the system database and include it in the plugin override.

The new function adds the override even if dex_idp is not enabled.
This override will only take effect if the user enables the
integration via user override. However, adding it to the plugin
override regardless poses no problem, so the function is being
called whether DEX integration is enabled or not.

The function get_value_from_application needed to be updated
to retrieve values from nested keys.

Test Plan:
PASS - Build all packages and STX-O tarball
PASS - Upload and apply tarball under 3 situations:
       1 - OIDC not applied and parameters not configured
       2 - OIDC not applied and parameters configured
       3 - OIDC applied
PASS - With parameters not configured and DEX integration enabled,
       Keystone plugin throws exception
PASS - issue_url retrieved in scenarios 2 and 3 regardless of
       DEX integration being enabled
PASS - App remove and delete

Story: 2011517
Task: 52959

Change-Id: If081fed9db9b4f751010581c15b469f95f1eb9d5
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>