Create sourced packages new folder structure
All packages that are sourced from upstream respositories
will be moved to the upstream/ folder.

This change creates a new folder named helm-charts/ under
the upstream directory, and moves the ingress-nginx package
to it. This change, also creates a new folder for the
openstack-helm package, but does not move the existing OSH
and OSH-I to it. This choice is part of the ongoing effort
to upversion the Openstack helm packages to the epoxy version.
The new OSH folder is created with only placeholder files, and
during the work to upversion OSH, it will be filled as to not
disrupt the build process during the upversion tasks.

The new structure is as follow:
openstack-armada-app:
|--- enhanced-policies
|--- openstack-helm
|--- openstack-helm-infra
|--- python3-k8sapp-openstack
|--- stx-openstack-helm-fluxcd
|--- upstream
     |--- helm-charts
          |--- ingress-nginx-helm
          |--- openstack-helm
     |--- openstack

As said in the paragraph above the original Oopenstack helm and Openstack Helm-Infra pacakages were not yet removed from the root dir of the repository, this will be done in future tasks when the process of upversion for the Openstack helm epoxy version is completed.

Test Plan:
PASS - Build all packages and STX-O tarball
PASS - Guarantee no error in the ingress-nginx-helm package buils
PASS - Upload and apply STX Openstack
PASs - Remove and Delete application

Story: 2011516
Task: 52984

Change-Id: Ic44a9d95df357d4dbf7d64d2b9cfe7881d8618fa
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>