Make Cinder storage-init flexible per enabled storage backends
This change introduces a patch that modifies the storage-init template.
It now checks if a backend is enabled (using the enabled_backends) to
dinamically create the K8s resources accordingly to the backend type.
For each enabled NetApp backend, a storage-init container is created
and as of this date, no action was implemented when we have a NetApp
backend. For Ceph backends, the behavior stays the same as before.

TEST PLAN:
PASS - Unit tests passed
PASS - build-pkgs -a -l openstack
PASS - build stx-openstack tarball
PASS - application upload, apply and remove
PASS - run 'openstack volume create' to check if the backends
are functional
PASS - storage backends tests as described below

Storage backends test:
The tests consists in applying some Ceph + NetApp combinations
so we can see check storage-init logs to check if the behavior is
what we expect.
For each backend, we should have a storage-init container. For the
NetApp containers, the logs should return a simple message:
"INFO: no action required to use <BACKEND>"

+-----------------+---------------------+--------------------+
| Backends        | No. of storage-init | Message displayed? |
|                 | containers          |                    |
+-----------------+---------------------+--------------------+
| Ceph            | 1                   | N/A                |
+-----------------+---------------------+--------------------+
| Ceph + NFS      | 2                   | Yes (1)            |
+-----------------+---------------------+--------------------+
| Ceph + iSCSI    | 2                   | Yes (1)            |
+-----------------+---------------------+--------------------+
| Ceph + NFS      | 3                   | Yes (2)            |
| + iSCSI         |                     |                    |
+-----------------+---------------------+--------------------+

Tests with NetApp only were not performed because the backup
related pods cannot apply without Ceph for now. This will be
addressed in the future.

Story: 2011281
Task: 53269

Change-Id: I97dea44622635afb74703e17ba0a2d4ea45f6630
Signed-off-by: Joao Fracarolli <joao.vicentinifracarolli@windriver.com>