Upversion cinderclient package
As part of the ongoing effort to upversion STX OpenStack to the Epoxy release, the cinderclient package needs to be updated.

This review updates the cinderclient package to version 9.7.0, which is one of the recommended versions for the OpenStack Epoxy release.

This review also adapts the deb patch 0001 to reflect changes made to the upstream package.

Test Plan:
PASS - Build all STX-O packages and wheels
PASS - Build STX-O tarball
PASS - Build STX docker images
PASS - Assert clients version on upgraded built images
BLOCKED - Apply STX-O overriding generated images

The apply failed due to an issue related to db_sync, which will be
handled in the Task: 53501

Depends-On: https://review.opendev.org/c/starlingx/tools/+/968174

Change-Id: I096ee7ad103173db40d7cadb8cdf7067bb2be78a
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>