Upversion OSH-I package to Epoxy release
As part of the ongoing effort to upversion STX-OpenStack
to the Epoxy release, the OSH-I package needs to be updated.

This change updates the metadata file by setting the commit
SHA to the commit that created the 2025.1.0 tag. It also
includes the adapted patches from an earlier review[1].
In addition, a new patch was created to adjust the Makefile
in the OSH repository to align it with the STX build environment.
Finally, the rules file was updated accordingly to reflect the
changes applied to the Makefile.

In future work, the OSH-I package will be merged into OSH.
For now, however, it will remain a separate package, even
though both currently point to the same commit in the
OSH repository.

Test Plan:
PASS - Build all packaged and STX-O tarball
PASS - OSH-I Helm charts are on the Epoxy release
PASS - Upload and Apply application on Host and Rook ceph systems
PASS - Sanity test on Simplex, Duplex and Standard systems

This review is part of a series of related changes required to
bring STX-OpenStack to the Epoxy release. The overall work
was split across multiple reviews to simplify the review process.
All reviews in this chain, along with the one in the fm-rest-api
chart[2], must be merged together to avoid breaking the application
build process.

[1] - https://review.opendev.org/c/starlingx/openstack-armada-app/+/968051
[2] - https://review.opendev.org/c/starlingx/helm-charts/+/969978

Story: 2011516
Task: 52972

Change-Id: I9fa35ed15835a456f6c1bafdea02f3c49db64dbc
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>