Update STX-Openstack FluxCD manifest
As part of the ongoing effort to upversion STX-OpenStack
to the Epoxy release, the FluxCD manifests must be updated
to reflect the changes made to the OSH and OSH-I packages.

This change updates the chart versions of all OSH Helm charts
in their HelmRelease files. It also updates the static overrides
for all charts sourced from the OSH repository to incorporate the changes made to the values files between the Caracal and Epoxy
releases.

Additionally, it was decided to update the version of the Helm
charts maintained by the StarlingX community to 2025.1.0, aligning
them with the chart versions provided by the OpenStack-Helm community.

Test Plan:
PASS - Build all packaged and STX-O tarball
PASS - Chart version in the Helmrelease updated to epoxy version
PASS - Upload and Apply application on Host and Rook ceph systems
PASS - Sanity test on Simplex, Duplex and Standard systems

This review is part of a series of related changes required to
bring STX-OpenStack to the Epoxy release. The overall work
was split across multiple reviews to simplify the review process.
All reviews in this chain, along with the one in the fm-rest-api
chart[Depends on section], must be merged together to avoid
breaking the application build process.

Depends-On: https://review.opendev.org/c/starlingx/helm-charts/+/969978

Story: 2011516
Task: 52972

Change-Id: Icda694cf30c172e0f39e45408a75d4c95658a2ab
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>