use `class:optional` in stx-openstack-helm-fluxcd
This modifies the metadata definition of the stx-openstack-helm-fluxcd
package to use `class: optional` instead of relying on the
`dependant_apps` definition to declare the dependency with
storage mechanisms such as rook-ceph.

By declaring the package as part of the `optional` class we want to
make sure the app is applied only after all storage class dependencies
are applied without explicitly declaring them as `dependant_apps`.

TEST PLAN:
[PASS] rebuild of stx-openstack-helm-fluxcd package
[PASS] WRO custom build on top of modified manifests
[PASS] WRO upload/apply/remove with rook-ceph only (no netapp)
[PASS] apply stx-openstack in a NetApp only deployment *
  * Some helm charts are still no compatible with NetApp only
    deployments making the apply operation to fail for reasons
    out of the scope of this change.

Story: 2011281
Task: 53575

Change-Id: I0a57d1b5a83537d67baedca6df580ed95aa1a901
Signed-off-by: Carlos Henrique <carlos.figueiredofelipe@windriver.com>